requirements.txt
streamlit
requests
langchain-text-splitters
duckduckgo-search
beautifulsoup4
qdrant-client
python-dotenv